requirements.txt
opencv-python
python-dotenv
ultralytics
torch
torchvision
